requirements.txt
requests
requests[socks]
pycryptodome
pynacl
fastapi
uvicorn
pillow